requirements.txt
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
seaborn==0.13.2
